requirements.txt
langchain-openai
langchain-community
python-dotenv
fastapi[standard]
